Move some files, combine summaryBlock and AnalysisDashboard
