+++ src/GlobalConfigParser.tsx
@@ -6,7 +6,7 @@ import ConfigSwitcher from './components/ConfigSwitcher';
 import { Shell } from './components/Shell';
 import { parseGlobalConfig } from './parser/parser';
 import { GlobalConfig, Nullable, ParsedStudyConfig } from './parser/types';
-import { AnalysisInterface } from './analysis/AnalysisInterface';
+import { AnalysisInterface } from './analysis/individualStudy/AnalysisInterface';
 import { PREFIX } from './utils/Prefix';
 import { ProtectedRoute } from './ProtectedRoute';
 import { Login } from './Login';
@@ -111,10 +111,6 @@ export function GlobalConfigParser() {
               />
               <Route
                 path="/analysis/stats/:studyId/:tab"
-                // loader={(params)=>{
-                //   console.log(params)
-                //   return false;
-                // }}
                 element={(
                   <ProtectedRoute paramToCheck="studyId" paramCallback={analysisProtectedCallback}>
                     <AnalysisInterface
