+++ src/analysis/dashboard/StudyCard.tsx
@@ -1,5 +1,5 @@
 import {
-  Badge, Box, Button, Card, Center, Text, Title, Container, Flex, Group, Popover,
+  Box, Button, Card, Center, Text, Title, Container, Flex, Group, Popover,
 } from '@mantine/core';
 import React, { useMemo, useState } from 'react';
 import { IconChartHistogram } from '@tabler/icons-react';
@@ -8,7 +8,7 @@ import { VegaLite } from 'react-vega';
 import { useDisclosure, useResizeObserver } from '@mantine/hooks';
 import { useNavigate } from 'react-router-dom';
 import { ParticipantData } from '../../storage/types';
-import { StoredAnswer, StudyConfig } from '../../parser/types';
+import { StoredAnswer } from '../../parser/types';
 import { DownloadButtons } from '../../components/downloader/DownloadButtons';
 import { ParticipantStatusBadges } from '../interface/ParticipantStatusBadges';
 
@@ -20,10 +20,7 @@ function isWithinRange(answers: Record<string, StoredAnswer>, rangeTime: [Date |
   return Math.min(...timeStamps) >= rangeTime[0].getTime() && Math.max(...timeStamps) <= rangeTime[1].getTime();
 }
 
-export function SummaryPanel(props: { studyId: string; allParticipants: ParticipantData[]; config: StudyConfig }) {
-  const {
-    studyId, allParticipants, config,
-  } = props;
+export function StudyCard({ studyId, allParticipants }: { studyId: string; allParticipants: ParticipantData[]; }) {
   const navigate = useNavigate();
   const [ref, dms] = useResizeObserver();
