+++ src/analysis/individualStudy/management/ManageAccordion.tsx
@@ -1,7 +1,7 @@
 import { Accordion, Container, Title } from '@mantine/core';
 import { DataManagementAccordionItem } from './DataManagementAccordionItem';
 import { RevisitModesAccordionItem } from './RevisitModesAccordionItem';
-import { useStorageEngine } from '../../storage/storageEngineHooks';
+import { useStorageEngine } from '../../../storage/storageEngineHooks';
 
 export default function ManageAccordion({ studyId, refresh }: { studyId: string, refresh: () => Promise<void> }) {
   const { storageEngine } = useStorageEngine();
